+++ app/Plugin/Admin/Site_Health.php
@@ -66,7 +66,7 @@ class Site_Health {
 	}
 
 	/**
-	 * Get list of endpoints the site health should use for our plugin.
+	 * Return list of endpoints the site health should use for our plugin.
 	 *
 	 * @return array<int,array<string,mixed>>
 	 */
@@ -79,16 +79,6 @@ class Site_Health {
 				'args'      => array(),
 			),
 		);
-
-		// add one check for each Personio URL.
-		foreach ( Personio_Accounts::get_instance()->get_personio_urls() as $personio_url ) {
-			$list[] = array(
-				'namespace' => 'personio/v1',
-				'route'     => '/url_availability_checks_' . md5( $personio_url ) . '/',
-				'callback'  => array( $this, 'url_availability_checks' ),
-				'args'      => array( array( 'personio_url' => $personio_url ) ),
-			);
-		}
 
 		/**
 		 * Filter the endpoints for Site Health this plugin is using.
@@ -190,60 +180,6 @@ class Site_Health {
 	}
 
 	/**
-	 * Check the Personio-URL availability.
-	 *
-	 * @param WP_REST_Request $request The request-object.
-	 *
-	 * @return array<string,mixed>
-	 */
-	public function url_availability_checks( WP_REST_Request $request ): array {
-		// get attributes to detect the requested Personio URL.
-		$args = $request->get_attributes();
-
-		// bail with error if no settings found.
-		if ( empty( $args ) || empty( $args['args'] ) || empty( $args['args'][0]['personio_url'] ) ) {
-			return array(
-				'label'       => __( 'Personio URL availability check', 'personio-integration-light' ),
-				'status'      => 'recommended',
-				'badge'       => array(
-					'label' => __( 'Personio Integration Light', 'personio-integration-light' ),
-					'color' => 'gray',
-				),
-				'description' => __( 'Missing Personio URL for check. Something is wrong with your plugin configuration.', 'personio-integration-light' ),
-				'action'      => '',
-				'test'        => 'personio_integration_rest_api_url_availability_check',
-			);
-		}
-
-		// get Personio-object for requested URL.
-		$personio_obj = new Personio( $args['args'][0]['personio_url'] );
-
-		// define default results.
-		$result = array(
-			'label'       => __( 'Personio URL availability Check', 'personio-integration-light' ),
-			'status'      => 'good',
-			'badge'       => array(
-				'label' => __( 'Personio Integration Light', 'personio-integration-light' ),
-				'color' => 'gray',
-			),
-			/* translators: %1$s and %2$s will be replaced by the Personio-URL */
-			'description' => sprintf( __( 'The Personio-URL <a href="%1$s" target="_blank">%2$s (opens new window)</a> is necessary to import new positions.<br><strong>All ok with the URL!</strong>', 'personio-integration-light' ), esc_url( $personio_obj->get_url() ), esc_url( $personio_obj->get_url() ) ),
-			'actions'     => '',
-			'test'        => 'personio_integration_rest_api_url_availability_check',
-		);
-
-		// request the HTTP-header of XML-API for the given Personio URL.
-		if ( ! PersonioIntegrationUrl::check_url( $personio_obj->get_url() ) ) {
-			$result['status'] = 'recommended';
-			/* translators: %1$s and %2$s will be replaced by the Personio-URL, %3$s will be replaced by the settings-URL, %4$s will be replaced by the URL to login on Personio */
-			$result['description'] = sprintf( __( 'The Personio-URL <a href="%1$s" target="_blank">%2$s (opens new window)</a> is not available for the import of positions!<br><strong>Please check if you have entered the correct URL <a href="%3$s">in the plugin-settings</a>.<br>Also check if you have enabled the XML-API in your <a href="%4$s" target="_blank">Personio-account (opens new window)</a> under Settings > Recruiting > Career Page > Activations.</strong>', 'personio-integration-light' ), esc_url( $personio_obj->get_url() ), esc_url( $personio_obj->get_url() ), esc_url( Helper::get_settings_url() ), esc_url( Helper::get_personio_login_url() ) );
-		}
-
-		// return result.
-		return $result;
-	}
-
-	/**
 	 * Add custom status-check for running cronjobs of our own plugin.
 	 * Only if Personio-URL is set.
 	 *
@@ -259,18 +195,7 @@ class Site_Health {
 			);
 		}
 
-		// one check for each Personio URL.
-		foreach ( Personio_Accounts::get_instance()->get_personio_urls() as $personio_url ) {
-			// get the md5.
-			$md5 = md5( $personio_url );
-
-			// add to list.
-			$statuses['async'][ 'personio_integration_url_availability_check_' . $md5 ] = array(
-				'label'    => __( 'Personio Integration URL availability check', 'personio-integration-light' ),
-				'test'     => rest_url( 'personio/v1/url_availability_checks_' . $md5 ),
-				'has_rest' => true,
-			);
-		}
+		// return the statuses.
 		return $statuses;
 	}
 }