+++ app/Plugin/Compatibilities.php
@@ -134,7 +134,7 @@ class Compatibilities {
 			'PersonioIntegrationLight\Plugin\Compatibilities\Elementor',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Everest_Forms',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Fluent_Forms',
-			'PersonioIntegrationLight\Plugin\Compatibilities\WPMultilang',
+			'PersonioIntegrationLight\Plugin\Compatibilities\Loco',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Nimble_Builder',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Oxygen',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Pdf_Generator_For_Wp',
@@ -149,6 +149,7 @@ class Compatibilities {
 			'PersonioIntegrationLight\Plugin\Compatibilities\Salient_WpBakery',
 			'PersonioIntegrationLight\Plugin\Compatibilities\WpBakery',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Wpml',
+			'PersonioIntegrationLight\Plugin\Compatibilities\WPMultilang',
 			'PersonioIntegrationLight\Plugin\Compatibilities\Yoast',
 		);