+++ templates/parts/part-grouptitle.php
@@ -5,7 +5,7 @@
  * @param array    $personio_attributes List of settings.
  * @param Position $position_obj       The positions object.
  *
- * @version: 4.0.0
+ * @version: 5.0.0
  * @package personio-integration-light
  */
 
@@ -22,6 +22,12 @@ if ( ! empty( $personio_attributes['groupby'] ) ) {
 	// output title if it has been changed during the loop.
 	if ( strcmp( $new_group_title, $group_title ) ) {
 		$group_title = $new_group_title;
+		if( $use_li ) {
+			echo '<li>';
+		}
 		echo '<h2>' . esc_html( $new_group_title ) . '</h2>';
+		if( $use_li ) {
+			echo '</li>';
+		}
 	}
 }